Make correct attribute data generation reusable
